* Add a first simple example notebook (mainly for paper purpose) * Refactor carla package output
