+++ carla/__init__.py
@@ -1,7 +1,6 @@
 # flake8: noqa
 
-from .evaluation import distances
-from .evaluation.nearest_neighbours import yNN
-from .evaluation.redundancy import redundancy
-from .evaluation.success_rate import success_rate
-from .evaluation.violations import constraint_violation
+from .data import Data, DataCatalog
+from .evaluation import Benchmark
+from .models import MLModel, MLModelCatalog
+from .recourse_methods import RecourseMethod
