+++ carla/data/__init__.py
@@ -1,1 +1,4 @@
 # flake8: noqa
+
+from .api import Data
+from .catalog import DataCatalog
