+++ carla/models/__init__.py
@@ -1,1 +1,4 @@
 # flake8: noqa
+
+from .api import MLModel
+from .catalog import MLModelCatalog
