+++ carla/recourse_methods/__init__.py
@@ -1,3 +1,4 @@
 # flake8: noqa
 
+from .api import RecourseMethod
 from .catalog import CEM, ActionableRecourse, Clue, Dice, Face, GrowingSpheres
