+++ test/test_yNN.py
@@ -1,5 +1,5 @@
-from carla import yNN
 from carla.data.catalog import DataCatalog
+from carla.evaluation import yNN
 from carla.models.catalog import MLModelCatalog
 from carla.models.negative_instances import predict_negative_instances
 from carla.recourse_methods.catalog.dice import Dice
